Mongodb cache (#191)
* Add test workflow

* Update file

* Update workflow

* Update workflow

* Add steps

* Try json functions

* Add line between jobs

* Update matrix

* Update cmd

* Update cmd

* Update cmd

* Update cmd

* Update workflow

* minro updates

* Update vars

* Debug output

* Update

* Reorg workflow

* Update

* Correct output name

* Update

* Update

* Add json.dumps to remove unicode

* Python list

* Remove last newline

* Initial commit

* Remove file

* Correct var name and add a break

* silence no_log warnings for update_password param (#192)

Silence the warning about no_log by setting no_log to False
on the update_password param of the mongodb_user module.

This takes the same approach as a postgresql module PR:
https://github.com/ansible/ansible/pull/68116

* Add workflow for cache tests

* Indentation

* Add default path

* Add pip install of pymongo

* Correct function call

* Install mongodb shell

* Add sudo

* Add output tests

* Remove output declaration

* Tidy

* Correct action path

* Rejig output

* Indentation

* stderr->stdout

* Comment out test

* Correct output path

* Correct output path

* Remove output header section

* Comment output problem step for now

* A mongo query checks

* Add assert back

* Update set-output

* Update step id

* Adjust set output

* Add debug step

* Fix syntax

* Add pipe

* stderr->stdout

* Escape newlines

* Update mongo cmds

* Add mongo checks

* Add notContains checks

* Add checks for _id_ index

* Remove test steps

+++ plugins/cache/mongodb.py
@@ -53,6 +53,8 @@ from ansible import constants as C
 from ansible.errors import AnsibleError
 from ansible.plugins.cache import BaseCacheModule
 from ansible.utils.display import Display
+from ansible.module_utils._text import to_native
+
 
 try:
     import pymongo
@@ -80,6 +82,22 @@ class CacheModule(BaseCacheModule):
         self._cache = {}
         self._managed_indexes = False
 
+    def _ttl_index_exists(self, collection):
+        '''
+        Returns true if an index named ttl exists
+        on the given collection.
+        '''
+        exists = False
+        try:
+            indexes = collection.list_indexes()
+            for index in indexes:
+                if index["name"] == "ttl":
+                    exists = True
+                    break
+        except pymongo.errors.OperationFailure as excep:
+            raise AnsibleError('Error checking MongoDB index: %s' % to_native(excep))
+        return exists
+
     def _manage_indexes(self, collection):
         '''
         This function manages indexes on the mongo collection.
@@ -96,10 +114,12 @@ class CacheModule(BaseCacheModule):
                 )
             except pymongo.errors.OperationFailure:
                 # We make it here when the fact_caching_timeout was set to a different value between runs
-                collection.drop_index('ttl')
+                if self._ttl_index_exists(collection):
+                    collection.drop_index('ttl')
                 return self._manage_indexes(collection)
         else:
-            collection.drop_index('ttl')
+            if self._ttl_index_exists(collection):
+                collection.drop_index('ttl')
 
     @contextmanager
     def _collection(self):
